Move http code into own file
